data-saver/requirements.txt
cassandra-driver
prettytable
python-dotenv
requests
pyopenssl
service-identity